+++ caddyhttp/rewrite/setup.go
@@ -37,8 +37,7 @@ func setup(c *caddy.Controller) error {
 }
 
 func rewriteParse(c *caddy.Controller) ([]httpserver.Config, error) {
-	var simpleRules []httpserver.Config
-	var regexpRules []httpserver.Config
+	var rules []httpserver.Config
 
 	for c.Next() {
 		var rule Rule
@@ -104,16 +103,15 @@ func rewriteParse(c *caddy.Controller) ([]httpserver.Config, error) {
 			if rule, err = NewComplexRule(base, pattern, to, status, ext, matcher); err != nil {
 				return nil, err
 			}
-			regexpRules = append(regexpRules, rule)
+			rules = append(rules, rule)
 
 		// the only unhandled case is 2 and above
 		default:
 			rule = NewSimpleRule(args[0], strings.Join(args[1:], " "))
-			simpleRules = append(simpleRules, rule)
+			rules = append(rules, rule)
 		}
 
 	}
 
-	// put simple rules in front to avoid regexp computation for them
-	return append(simpleRules, regexpRules...), nil
+	return rules, nil
 }
