+++ caddyhttp/rewrite/to.go
@@ -12,8 +12,7 @@ import (
 
 // To attempts rewrite. It attempts to rewrite to first valid path
 // or the last path if none of the paths are valid.
-// Returns true if rewrite is successful and false otherwise.
-func To(fs http.FileSystem, r *http.Request, to string, replacer httpserver.Replacer) Result {
+func To(fs http.FileSystem, r *http.Request, to string, replacer httpserver.Replacer) (result Result) {
 	tos := strings.Fields(to)
 
 	// try each rewrite paths
@@ -45,7 +44,7 @@ func To(fs http.FileSystem, r *http.Request, to string, replacer httpserver.Repl
 		// Let the user know we got here. Rewrite is expected but
 		// the resulting url is invalid.
 		log.Printf("[ERROR] rewrite: resulting path '%v' is invalid. error: %v", t, err)
-		return RewriteIgnored
+		return
 	}
 
 	// take note of this rewrite for internal use by fastcgi
